feat: SELinux policy to prevent moving or deleting .config
This enhances the bash environment lockdown with an SELinux policy
module that prevents the directory `$HOME/.config` from being moved,
deleted, or mounted over.

The restrictions on `.config` are controlled by an SELinux boolean,
`homedir_lockdown_enabled`, that defaults to off. When it is on,
relabeling to and from `config_locked_home_t` is also disallowed. The
bash environment lockdown changes `.config` to have this type and then
toggles the boolean on.

Also added a check to the audit script for this enhancement, informing
the user if the bash environment lockdown is on but `.config` doesn't
have the `config_locked_home_t` type.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>