ci: include commit list in PR body for update-modules.yml
To make review of BlueBuild module update PRs easier, have the workflow
pull the list of commits to the BlueBuild modules repo that were made
since the previous module update, and include those in a list in the PR
body.

Also skip making a PR entirely if there are no new commits.